pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "davinci-resolve-cli"
version = "0.1.0"
description = "A CLI for DaVinci Resolve 18+ — project, media, render, and timeline control for humans and AI agents."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "钟超", email = "zhongchao.ustc@gmail.com" }]
keywords = ["davinci-resolve", "cli", "video-editing", "ai-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "typer >= 0.12.0",
    "rich >= 13.7.0",
    "pyyaml >= 6.0",
    "jsonschema >= 4.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest >= 8.0",
    "pytest-cov >= 5.0",
    "ruff >= 0.4.0",
    "mypy >= 1.10",
]

[project.scripts]
dvr = "dvr.cli:main_entry"

[project.urls]
Homepage = "https://github.com/poechant/davinci-resolve-cli"
Repository = "https://github.com/poechant/davinci-resolve-cli"
Issues = "https://github.com/poechant/davinci-resolve-cli/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/dvr"]

[tool.hatch.build.targets.wheel.shared-data]
"SKILL.md" = "share/dvr/SKILL.md"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: requires a running DaVinci Resolve 18+ instance",
]
addopts = "-ra -m 'not integration'"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
